Action.cancel이 isCancellable 조건을 무시하지 않도록 수정, 개별 스킬에서 취소 조건 확인

+++ src/main/java/com/dace/dmgr/combat/combatant/inferno/action/InfernoA2.java
@@ -152,7 +152,7 @@ public final class InfernoA2 extends ActiveSkill implements HasBonusScore {
                 target.getStatusEffectModule().apply(Grounding.getInstance(), combatUser, Timespan.ofTicks(10));
 
                 if (target instanceof CombatUser) {
-                    combatUser.addScore("적 고정", (double) (InfernoA2Info.EFFECT_SCORE_PER_SECOND * 4) / 20);
+                    combatUser.addScore("적 고정", InfernoA2Info.EFFECT_SCORE_PER_SECOND * 4 / 20.0);
                     bonusScoreModule.addTarget((CombatUser) target, Timespan.ofTicks(10));
                 }
             }
